requirements.txt
python-telegram-bot
sqlalchemy
databases[postgresql]
databases[mysql]
databases[sqlite]
